Problem: BigchainDB class should be part of project root
Solution: Removed the /tendermint directory and moved its contents to project root
