+++ bigchaindb/__init__.py
@@ -84,5 +84,6 @@ config = {
 # the user wants to reconfigure the node. Check ``bigchaindb.config_utils``
 # for more info.
 _config = copy.deepcopy(config)
-from bigchaindb.tendermint import BigchainDB  # noqa
+from bigchaindb.lib import BigchainDB  # noqa
+from bigchaindb.core import App
 from bigchaindb.version import __version__  # noqa
