+++ bigchaindb/fastquery.py
