+++ bigchaindb/lib.py
@@ -16,13 +16,12 @@ except ImportError:
 import requests
 
 import bigchaindb
-from bigchaindb import backend, config_utils
+from bigchaindb import backend, config_utils, fastquery
 from bigchaindb.models import Transaction
 from bigchaindb.common.exceptions import (SchemaValidationError,
                                           ValidationError,
                                           DoubleSpend)
-from bigchaindb.tendermint.utils import encode_transaction, merkleroot
-from bigchaindb.tendermint import fastquery
+from bigchaindb.tendermint_utils import encode_transaction, merkleroot
 from bigchaindb import exceptions as core_exceptions
 from bigchaindb.consensus import BaseConsensusRules
