+++ docs/server/generate_http_server_api_documentation.py
@@ -5,8 +5,7 @@ import os
 import os.path
 
 from bigchaindb.common.transaction import Transaction, Input, TransactionLink
-from bigchaindb.tendermint import BigchainDB
-from bigchaindb.tendermint import lib
+from bigchaindb import lib
 from bigchaindb.web import server
