--- docs/server/source/appendices/tendermint-integration.rst
@@ -1,26 +1,0 @@
-######################
-Tendermint Integration
-######################
-
-
-.. automodule:: bigchaindb.tendermint
-    :special-members: __init__
-
-.. automodule:: bigchaindb.tendermint.lib
-    :special-members: __init__
-    :noindex:
-
-.. automodule:: bigchaindb.tendermint.core
-    :special-members: __init__
-
-.. automodule:: bigchaindb.tendermint.event_stream
-    :special-members: __init__
-
-.. automodule:: bigchaindb.tendermint.fastquery
-    :special-members: __init__
-
-.. automodule:: bigchaindb.tendermint.commands
-    :special-members: __init__
-
-.. automodule:: bigchaindb.tendermint.utils
-    :special-members: __init__
