+++ tests/tendermint/test_utils.py
@@ -12,7 +12,7 @@ pytestmark = pytest.mark.tendermint
 
 
 def test_encode_decode_transaction(b):
-    from bigchaindb.tendermint.utils import (encode_transaction,
+    from bigchaindb.tendermint_utils import (encode_transaction,
                                              decode_transaction)
 
     asset = {
@@ -30,7 +30,7 @@ def test_encode_decode_transaction(b):
 
 
 def test_calculate_hash_no_key(b):
-    from bigchaindb.tendermint.utils import calculate_hash
+    from bigchaindb.tendermint_utils import calculate_hash
 
     # pass an empty list
     assert calculate_hash([]) == ''
@@ -38,7 +38,7 @@ def test_calculate_hash_no_key(b):
 
 # TODO test for the case of an empty list of hashes, and possibly other cases.
 def test_merkleroot():
-    from bigchaindb.tendermint.utils import merkleroot
+    from bigchaindb.tendermint_utils import merkleroot
     hashes = [sha3_256(i.encode()).digest() for i in 'abc']
     assert merkleroot(hashes) == (
         '78c7c394d3158c218916b7ae0ebdea502e0f4e85c08e3b371e3dfd824d389fa3')
@@ -54,14 +54,14 @@ SAMPLE_PUBLIC_KEY = {
 
 
 def test_convert_base64_public_key_to_address():
-    from bigchaindb.tendermint.utils import public_key64_to_address
+    from bigchaindb.tendermint_utils import public_key64_to_address
 
     address = public_key64_to_address(SAMPLE_PUBLIC_KEY['pub_key']['value'])
     assert address == SAMPLE_PUBLIC_KEY['address']
 
 
 def test_public_key_encoding_decoding():
-    from bigchaindb.tendermint.utils import (public_key_from_base64,
+    from bigchaindb.tendermint_utils import (public_key_from_base64,
                                              public_key_to_base64)
 
     public_key = public_key_from_base64(SAMPLE_PUBLIC_KEY['pub_key']['value'])
