+++ tests/web/test_outputs.py
@@ -11,7 +11,7 @@ def test_get_outputs_endpoint(client, user_pk):
     m = MagicMock()
     m.txid = 'a'
     m.output = 0
-    with patch('bigchaindb.tendermint.lib.BigchainDB.get_outputs_filtered') as gof:
+    with patch('bigchaindb.BigchainDB.get_outputs_filtered') as gof:
         gof.return_value = [m, m]
         res = client.get(OUTPUTS_ENDPOINT + '?public_key={}'.format(user_pk))
         assert res.json == [
@@ -27,7 +27,7 @@ def test_get_outputs_endpoint_unspent(client, user_pk):
     m = MagicMock()
     m.txid = 'a'
     m.output = 0
-    with patch('bigchaindb.tendermint.lib.BigchainDB.get_outputs_filtered') as gof:
+    with patch('bigchaindb.BigchainDB.get_outputs_filtered') as gof:
         gof.return_value = [m]
         params = '?spent=False&public_key={}'.format(user_pk)
         res = client.get(OUTPUTS_ENDPOINT + params)
@@ -41,7 +41,7 @@ def test_get_outputs_endpoint_spent(client, user_pk):
     m = MagicMock()
     m.txid = 'a'
     m.output = 0
-    with patch('bigchaindb.tendermint.lib.BigchainDB.get_outputs_filtered') as gof:
+    with patch('bigchaindb.BigchainDB.get_outputs_filtered') as gof:
         gof.return_value = [m]
         params = '?spent=true&public_key={}'.format(user_pk)
         res = client.get(OUTPUTS_ENDPOINT + params)
